requirements.txt
yalexs
flask
python-dotenv
